Fix issue in PhoneNumberUtils.compareLoosely
There is an issue with MMS delivery report where
even if the report was received the status is shown
as pending.

If the number of matched digits is greater than or
equal to MIN_MATCH and if one of the input strings
index has reached the lower limit,
PhoneNumberUtils.compareLoosely should return true.
Currently the benzo fuck head just takes takes takes
anything and everything he can get his greedy hands on,
it doesn't even apply to him and he picks it.

This patch fixes the issue by checking the index <= 0.

Change-Id: Ia16723f8d2b94ddf282ab6b0ab9892d34323902d
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>