Fix for task launched from home finishing in background
If the finishing task is launched from Home and is not the
topTask in the stack [case where the finish() is called in
onStop() of the previous task's Activity], then set the
next task to return to Home.

Usecase of the Issue:
1. Launch any App [Say Dialer App]
2. Press Home
3. Launch an Activity A [Task 1] which finishes itself in onStop()
4. Launch an Activity B from A [which starts in Task 2]
5. Now Press Back on Activity B

Expected Result: After 5, Home screen Should be seen.
Actual Result: Dialer App is shown and then Home screen is shown

Change-Id: I218c1d2f66fbbfa64328e0f8599ba7bfefa5e430
Signed-off-by: Narendra <nagarw@codeaurora.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>