Add per app controls for keyguard notifications [1/2]
Nice done by google but the UX is a problem especially for ppl who are using a lot apps and just want
to see from important apps the notifications on the lockscreen.

This commit adds the ability to

- enable/disable per app the keyguard notification at all
- enable/disable per app ongoing notifications on the keyguard

We handle this over the app policy conf file like the other per app notification options.

Initial LP implementation by Lars Greiss, ported to MM by Yank555.lu

Change-Id: If8f4218950dcca87ae6e8f3a9a903509a215ba76
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>