Keep stack to front for not-running task comes to front.
When user starts a task which is not running by choosing it on
RecentApp, Home stack come to front instead application stack since
there is no running task on the stack. Eventually the task cat not come
up to front but it starts on background.

Explictly keep stack to front in case of no task is in the stack, just
like Lollipop does.

Change-Id: I67093b3bd66be1ccf785b66fcc1300d26804b541
Signed-off-by: mydongistiny <jaysonedson@gmail.com>