Fonts: add sans-serif fallback fonts first
This fixes an issue where the non line breaking hyphen U+2011 being
displayed as and underscore due to the glyph being found in one of
the fallback fonts.  Adding the sans-serif(Roboto) fonts as fallbacks
first allows the glyph to be found in this family before moving on
to the other fallback fonts.

Change-Id: Ie6ac7e8c838f3a2b864ff25bc9569a8f37beffb9
TICKET: CYNGNOS-1855
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>