JobScheduler: Deadline constraint will be waked up by mistake
If a job only has deadline constraint, then isConstraintsSatisfied()
and isReady() will be always true, which results the job always be ready.
So when a new job is added or some constraints happened for other jobs,
the job of "only has deadline constraint" will be added into pending
job list and be fired to run, but actually the elasped time is not up
to deadline.

Bug https://code.google.com/p/android/issues/detail?id=206401

Change-Id: I5d4e8730b4ce26d76a0edde6d876d5afb1b4eebe